third_party/WebKit/LayoutTests/bluetooth/idl-NavigatorBluetooth.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script>
test(function() {
  assert_true('bluetooth' in navigator,
              'navigator.bluetooth exists.');
}, 'navigator.bluetooth IDL test');
</script>
